requirements.txt
fastapi[standard]
sqlalchemy
fastapi[standard]
pyjwt
python-multipart
passlib[bcrypt]